+++ .changeset/tough-beans-unite.md
@@ -1,0 +1,23 @@
+---
+'@finos/legend-application-pure-ide-deployment': patch
+'@finos/legend-application-taxonomy-deployment': patch
+'@finos/legend-application-taxonomy-bootstrap': patch
+'@finos/legend-application-studio-deployment': patch
+'@finos/legend-application-query-deployment': patch
+'@finos/legend-application-query-bootstrap': patch
+'@finos/legend-extension-dsl-data-space': patch
+'@finos/legend-extension-dsl-diagram': patch
+'@finos/legend-extension-dsl-service': patch
+'@finos/legend-application-pure-ide': patch
+'@finos/legend-application-taxonomy': patch
+'@finos/legend-application-studio': patch
+'@finos/legend-application-query': patch
+'@finos/legend-query-builder': patch
+'@finos/legend-dev-utils': patch
+'@finos/stylelint-config-legend-studio': patch
+'@finos/eslint-plugin-legend-studio': patch
+'@finos/legend-shared': patch
+'@finos/babel-preset-legend-studio': patch
+'@finos/legend-lego': patch
+'@finos/legend-art': patch
+---
