--- packages/legend-extension-dsl-data-space/src/components/DSL_DataSpace_LegendApplicationPlugin.ts
@@ -1,54 +1,0 @@
-/**
- * Copyright (c) 2020-present, Goldman Sachs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import {
-  LegendApplicationPlugin,
-  type KeyedCommandConfigEntry,
-  type LegendApplicationPluginManager,
-  type LegendApplicationSetup,
-  collectKeyedCommandConfigEntriesFromConfig,
-} from '@finos/legend-application';
-import packageJson from '../../package.json';
-import { configureDataGridComponent } from '@finos/legend-lego/data-grid';
-import { DSL_DATA_SPACE_LEGEND_APPLICATION_COMMAND_CONFIG } from '../__lib__/DSL_DataSpace_LegendApplicationCommand.js';
-
-export class DSL_DataSpace_LegendApplicationPlugin extends LegendApplicationPlugin {
-  static NAME = packageJson.extensions.applicationPlugin;
-
-  constructor() {
-    super(DSL_DataSpace_LegendApplicationPlugin.NAME, packageJson.version);
-  }
-
-  install(
-    pluginManager: LegendApplicationPluginManager<LegendApplicationPlugin>,
-  ): void {
-    pluginManager.registerApplicationPlugin(this);
-  }
-
-  override getExtraKeyedCommandConfigEntries(): KeyedCommandConfigEntry[] {
-    return collectKeyedCommandConfigEntriesFromConfig(
-      DSL_DATA_SPACE_LEGEND_APPLICATION_COMMAND_CONFIG,
-    );
-  }
-
-  override getExtraApplicationSetups(): LegendApplicationSetup[] {
-    return [
-      async (applicationStore) => {
-        configureDataGridComponent();
-      },
-    ];
-  }
-}
