+++ packages/legend-extension-dsl-data-space/src/components/DSL_DataSpace_LegendApplicationPlugin.tsx
@@ -16,7 +16,10 @@
 
 import {
   LegendApplicationPlugin,
+  type LegendApplicationSetup,
   type LegendApplicationPluginManager,
+  collectKeyedCommandConfigEntriesFromConfig,
+  type KeyedCommandConfigEntry,
 } from '@finos/legend-application';
 import packageJson from '../../package.json';
 import type {
@@ -24,11 +27,13 @@ import type {
   QueryBuilder_LegendApplicationPlugin_Extension,
 } from '@finos/legend-query-builder';
 import { DataSpaceQueryBuilderState } from '../stores/query/DataSpaceQueryBuilderState.js';
-import type { QuerySearchSpecification } from '@finos/legend-graph';
 import {
   createQueryClassTaggedValue,
   createQueryDataSpaceTaggedValue,
 } from '../stores/query/DataSpaceQueryCreatorStore.js';
+import { DSL_DATA_SPACE_LEGEND_APPLICATION_COMMAND_CONFIG } from '../__lib__/DSL_DataSpace_LegendApplicationCommand.js';
+import type { QuerySearchSpecification } from '@finos/legend-graph';
+import { configureDataGridComponent } from '@finos/legend-lego/data-grid';
 
 export class DSL_DataSpace_LegendApplicationPlugin
   extends LegendApplicationPlugin
@@ -44,6 +49,20 @@ export class DSL_DataSpace_LegendApplicationPlugin
     pluginManager: LegendApplicationPluginManager<LegendApplicationPlugin>,
   ): void {
     pluginManager.registerApplicationPlugin(this);
+  }
+
+  override getExtraKeyedCommandConfigEntries(): KeyedCommandConfigEntry[] {
+    return collectKeyedCommandConfigEntriesFromConfig(
+      DSL_DATA_SPACE_LEGEND_APPLICATION_COMMAND_CONFIG,
+    );
+  }
+
+  override getExtraApplicationSetups(): LegendApplicationSetup[] {
+    return [
+      async (applicationStore) => {
+        configureDataGridComponent();
+      },
+    ];
   }
 
   getExtraLoadQueryFilterOptions(): LoadQueryFilterOption[] {
